Remove internal imports from parser examples
